scripts: don't build the index when pages haven't been changed, fix deploy.sh

+++ scripts/deploy.sh
@@ -30,10 +30,10 @@ function initialize {
 
 function upload_assets {
   git clone --quiet --depth 1 git@github.com:${SITE_REPO_SLUG}.git "$SITE_HOME"
-  mv -f "$TLDR_ARCHIVE" "$SITE_HOME/assets/"
+  [[ -f "$TLDR_ARCHIVE" ]] && mv -f "$TLDR_ARCHIVE" "$SITE_HOME/assets/"
   find "$TLDR_LANG_ARCHIVES_DIRECTORY" -maxdepth 1 -name "*.zip" -exec mv -f {} "$SITE_HOME/assets/" \;
   rm -rf "$TLDR_LANG_ARCHIVES_DIRECTORY"
-  cp -f "$TLDRHOME/index.json" "$SITE_HOME/assets/"
+  [[ -f "$TLDRHOME/index.json" ]] && cp -f "$TLDRHOME/index.json" "$SITE_HOME/assets/"
   find "$TLDR_PDF_FILES_DIRECTORY" -maxdepth 1 -name "*.pdf" -exec mv -f {} "$SITE_HOME/assets/" \;
   rm -rf "$TLDR_PDF_FILES_DIRECTORY"
